# apps/api/src/connections/connections.controller.ts
import {
  Body,
  Controller,
  Delete,
  Get,
  Logger,
  Param,
  Post,
  Query,
  Req,
  Res,
} from "@nestjs/common";
import { ApiBearerAuth, ApiOperation, ApiTags } from "@nestjs/swagger";
import { Throttle } from "@nestjs/throttler";
import {
  discordConnectSchema,
  telegramConnectSchema,
  xConnectSchema,
  type DiscordConnectInput,
  type TelegramConnectInput,
  type XConnectInput,
} from "@godeye/shared";
import type { Request, Response } from "express";
import { CurrentAuth } from "../common/current-auth.decorator";
import { AccessTokenPayload, JwtAuthGuard } from "../common/jwt-auth.guard";
import { env } from "../common/env";
import { ZodPipe } from "../common/zod.pipe";
import { ConnectionsService } from "./connections.service";
import { Public } from "../common/public.decorator";
import { MinRole } from "../common/roles.guard";

@ApiTags("connections")
@Controller("connections")
export class ConnectionsController {
  private readonly logger = new Logger(ConnectionsController.name);

  constructor(private readonly connections: ConnectionsService) {}

  @Get()
  @MinRole("VIEWER")
  @ApiBearerAuth()
  list(@CurrentAuth() auth: AccessTokenPayload) {
    return this.connections.list(auth.orgId);
  }

  @Delete(":id")
  @MinRole("ADMIN")
  @ApiBearerAuth()
  remove(@CurrentAuth() auth: AccessTokenPayload, @Param("id") id: string) {
    return this.connections.remove(auth.orgId, id, auth.sub);
  }

  @Post("telegram")
  @MinRole("ADMIN")
  @ApiBearerAuth()
  @Throttle({ default: { limit: 10, ttl: 60_000 } })
  @ApiOperation({ summary: "Connect a Telegram bot + channel" })
  connectTelegram(
    @CurrentAuth() auth: AccessTokenPayload,
    @Body(new ZodPipe(telegramConnectSchema)) body: TelegramConnectInput,
  ) {
    return this.connections.connectTelegram(auth.orgId, auth.sub, body);
  }

  @Post("discord")
  @MinRole("ADMIN")
  @ApiBearerAuth()
  @Throttle({ default: { limit: 10, ttl: 60_000 } })
  @ApiOperation({ summary: "Connect a Discord bot + channel" })
  connectDiscord(
    @CurrentAuth() auth: AccessTokenPayload,
    @Body(new ZodPipe(discordConnectSchema)) body: DiscordConnectInput,
  ) {
    return this.connections.connectDiscord(auth.orgId, auth.sub, body);
  }

  @Get("reddit/authorize")
  @MinRole("ADMIN")
  @ApiBearerAuth()
  @ApiOperation({ summary: "Get the Reddit OAuth dialog URL (click-to-connect)" })
  redditAuthorize(
    @CurrentAuth() auth: AccessTokenPayload,
    @Res({ passthrough: true }) res: Response,
  ) {
    return this.connections.redditAuthorize(res, auth.orgId, auth.sub);
  }

  @Get("reddit/callback")
  @Public()
  @ApiOperation({ summary: "OAuth redirect target for Reddit, do not call directly" })
  async redditCallback(
    @Query("code") code: string,
    @Query("state") state: string,
    @Query("error") error: string | undefined,
    @Req() req: Request,
    @Res() res: Response,
  ) {
    const base = `${env.webUrl}/connections`;
    if (error || !code || !state) {
      return res.redirect(
        `${base}?error=${encodeURIComponent(error ?? "Reddit authorization failed")}`,
      );
    }
    try {
      await this.connections.redditCallback(req, res, code, state);
      return res.redirect(`${base}?connected=reddit&count=1`);
    } catch (e) {
      const message = e instanceof Error ? e.message : "Reddit connection failed";
      return res.redirect(`${base}?error=${encodeURIComponent(message)}`);
    }
  }

  @Get("x/authorize")
  @MinRole("ADMIN")
  @ApiBearerAuth()
  @Throttle({ default: { limit: 10, ttl: 60_000 } })
  @ApiOperation({ summary: "Get the X OAuth dialog URL (click-to-connect)" })
  xAuthorize(@CurrentAuth() auth: AccessTokenPayload) {
    return this.connections.xAuthorize(auth.orgId, auth.sub);
  }

  @Get("x/callback")
  @Public()
  @ApiOperation({ summary: "OAuth redirect target for X, do not call directly" })
  async xCallback(
    @Query("oauth_token") oauthToken: string,
    @Query("oauth_verifier") verifier: string,
    @Query("denied") denied: string | undefined,
    @Res() res: Response,
  ) {
    const base = `${env.webUrl}/connections`;
    // X sends ?denied=<token> rather than an error parameter when someone
    // presses Cancel, and no verifier with it.
    if (denied) return res.redirect(`${base}?error=${encodeURIComponent("X authorization cancelled")}`);
    if (!oauthToken || !verifier) {
      return res.redirect(`${base}?error=${encodeURIComponent("X authorization failed")}`);
    }
    try {
      await this.connections.xCallback(oauthToken, verifier);
      return res.redirect(`${base}?connected=x&count=1`);
    } catch (e) {
      const message = e instanceof Error ? e.message : "X connection failed";
      return res.redirect(`${base}?error=${encodeURIComponent(message)}`);
    }
  }

  /**
   * What this creator's TikTok account allows.
   *
   * VIEWER, not EDITOR: the composer calls it to render the publish options,
   * and it returns the account's own settings, nothing a viewer could not
   * already see in TikTok.
   */
  @Get(":id/tiktok/creator-info")
  @MinRole("VIEWER")
  @ApiBearerAuth()
  @Throttle({ default: { limit: 30, ttl: 60_000 } })
  @ApiOperation({
    summary: "The privacy levels and interaction settings this TikTok account permits",
  })
  tiktokCreatorInfo(@CurrentAuth() auth: AccessTokenPayload, @Param("id") id: string) {
    return this.connections.tiktokCreatorInfo(auth.orgId, id);
  }

  @Post("x")
  @MinRole("ADMIN")
  @ApiBearerAuth()
  @Throttle({ default: { limit: 10, ttl: 60_000 } })
  @ApiOperation({
    summary:
      "Connect X with an access token pair generated by hand, the redirect flow at x/authorize is the normal path",
  })
  connectX(
    @CurrentAuth() auth: AccessTokenPayload,
    @Body(new ZodPipe(xConnectSchema)) body: XConnectInput,
  ) {
    return this.connections.connectX(auth.orgId, auth.sub, body);
  }

  @Get("linkedin/authorize")
  @MinRole("ADMIN")
  @ApiBearerAuth()
  @ApiOperation({ summary: "Get the LinkedIn OAuth dialog URL" })
  linkedinAuthorize(
    @CurrentAuth() auth: AccessTokenPayload,
    @Res({ passthrough: true }) res: Response,
  ) {
    return this.connections.linkedinAuthorize(res, auth.orgId, auth.sub);
  }

  @Get("linkedin/callback")
  @Public()
  @ApiOperation({ summary: "OAuth redirect target for LinkedIn, do not call directly" })
  async linkedinCallback(
    @Query("code") code: string,
    @Query("state") state: string,
    @Query("error_description") errorDescription: string | undefined,
    @Req() req: Request,
    @Res() res: Response,
  ) {
    const base = `${env.webUrl}/connections`;
    if (errorDescription || !code || !state) {
      return res.redirect(
        `${base}?error=${encodeURIComponent(errorDescription ?? "LinkedIn authorization failed")}`,
      );
    }
    try {
      await this.connections.linkedinCallback(req, res, code, state);
      return res.redirect(`${base}?connected=linkedin&count=1`);
    } catch (e) {
      const message = e instanceof Error ? e.message : "LinkedIn connection failed";
      return res.redirect(`${base}?error=${encodeURIComponent(message)}`);
    }
  }

  @Get("tiktok/authorize")
  @MinRole("ADMIN")
  @ApiBearerAuth()
  @ApiOperation({ summary: "Get the TikTok OAuth dialog URL" })
  tiktokAuthorize(
    @CurrentAuth() auth: AccessTokenPayload,
    @Res({ passthrough: true }) res: Response,
  ) {
    return this.connections.tiktokAuthorize(res, auth.orgId, auth.sub);
  }

  @Get("tiktok/callback")
  @Public()
  @ApiOperation({ summary: "OAuth redirect target for TikTok, do not call directly" })
  async tiktokCallback(
    @Query("code") code: string,
    @Query("state") state: string,
    @Query("error_description") errorDescription: string | undefined,
    @Req() req: Request,
    @Res() res: Response,
  ) {
    const base = `${env.webUrl}/connections`;
    if (errorDescription || !code || !state) {
      return res.redirect(
        `${base}?error=${encodeURIComponent(errorDescription ?? "TikTok authorization failed")}`,
      );
    }
    try {
      const result = await this.connections.tiktokCallback(req, res, code, state);
      return res.redirect(`${base}?connected=tiktok&count=${result.connected}`);
    } catch (e) {
      const message = e instanceof Error ? e.message : "TikTok connection failed";
      this.logger.error(
        `TikTok callback failed: ${message}`,
        e instanceof Error ? e.stack : undefined,
      );
      return res.redirect(`${base}?error=${encodeURIComponent(message)}`);
    }
  }

  @Get("instagram/authorize")
  @MinRole("ADMIN")
  @ApiBearerAuth()
  @ApiOperation({
    summary: "Get the Instagram OAuth dialog URL (no Facebook Page required)",
  })
  instagramAuthorize(
    @CurrentAuth() auth: AccessTokenPayload,
    @Res({ passthrough: true }) res: Response,
  ) {
    return this.connections.instagramAuthorize(res, auth.orgId, auth.sub);
  }

  @Get("instagram/callback")
  @Public()
  @ApiOperation({ summary: "OAuth redirect target for Instagram, do not call directly" })
  async instagramCallback(
    @Query("code") code: string,
    @Query("state") state: string,
    @Query("error_description") errorDescription: string | undefined,
    @Req() req: Request,
    @Res() res: Response,
  ) {
    const base = `${env.webUrl}/connections`;
    if (errorDescription || !code || !state) {
      return res.redirect(
        `${base}?error=${encodeURIComponent(errorDescription ?? "Instagram authorization failed")}`,
      );
    }
    try {
      const result = await this.connections.instagramCallback(req, res, code, state);
      return res.redirect(`${base}?connected=instagram&count=${result.connected}`);
    } catch (e) {
      const message = e instanceof Error ? e.message : "Instagram connection failed";
      this.logger.error(
        `Instagram callback failed: ${message}`,
        e instanceof Error ? e.stack : undefined,
      );
      return res.redirect(`${base}?error=${encodeURIComponent(message)}`);
    }
  }

  @Get("meta/authorize")
  @MinRole("ADMIN")
  @ApiBearerAuth()
  @ApiOperation({ summary: "Get the Facebook OAuth dialog URL (Facebook Pages + Instagram)" })
  metaAuthorize(
    @CurrentAuth() auth: AccessTokenPayload,
    @Res({ passthrough: true }) res: Response,
  ) {
    return this.connections.metaAuthorize(res, auth.orgId, auth.sub);
  }

  @Get("meta/callback")
  @Public()
  @ApiOperation({ summary: "OAuth redirect target for Meta, do not call directly" })
  async metaCallback(
    @Query("code") code: string,
    @Query("state") state: string,
    @Query("error_description") errorDescription: string | undefined,
    @Req() req: Request,
    @Res() res: Response,
  ) {
    const base = `${env.webUrl}/connections`;
    if (errorDescription || !code || !state) {
      return res.redirect(`${base}?error=${encodeURIComponent(errorDescription ?? "Meta authorization failed")}`);
    }
    try {
      const result = await this.connections.metaCallback(req, res, code, state);
      return res.redirect(`${base}?connected=meta&count=${result.connected}`);
    } catch (e) {
      // Log it: the message otherwise only exists in the redirect URL, which the
      // user never reads if the browser bounces them to /login on the way.
      const message = e instanceof Error ? e.message : "Meta connection failed";
      this.logger.error(`Meta callback failed: ${message}`, e instanceof Error ? e.stack : undefined);
      return res.redirect(`${base}?error=${encodeURIComponent(message)}`);
    }
  }
}