# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/connections": {
      "get": {
        "operationId": "ConnectionsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "connections"
        ]
      }
    },
    "/connections/{id}": {
      "delete": {
        "operationId": "ConnectionsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "connections"
        ]
      }
    },
    "/connections/telegram": {
      "post": {
        "operationId": "ConnectionsController_connectTelegram",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Connect a Telegram bot + channel",
        "tags": [
          "connections"
        ]
      }
    },
    "/connections/discord": {
      "post": {
        "operationId": "ConnectionsController_connectDiscord",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Connect a Discord bot + channel",
        "tags": [
          "connections"
        ]
      }
    },
    "/connections/reddit/authorize": {
      "get": {
        "operationId": "ConnectionsController_redditAuthorize",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the Reddit OAuth dialog URL (click-to-connect)",
        "tags": [
          "connections"
        ]
      }
    },
    "/connections/reddit/callback": {
      "get": {
        "operationId": "ConnectionsController_redditCallback",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "OAuth redirect target for Reddit, do not call directly",
        "tags": [
          "connections"
        ]
      }
    },
    "/connections/x/authorize": {
      "get": {
        "operationId": "ConnectionsController_xAuthorize",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the X OAuth dialog URL (click-to-connect)",
        "tags": [
          "connections"
        ]
      }
    },
    "/connections/x/callback": {
      "get": {
        "operationId": "ConnectionsController_xCallback",
        "parameters": [
          {
            "name": "oauth_token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "oauth_verifier",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "OAuth redirect target for X, do not call directly",
        "tags": [
          "connections"
        ]
      }
    },
    "/connections/{id}/tiktok/creator-info": {
      "get": {
        "operationId": "ConnectionsController_tiktokCreatorInfo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The privacy levels and interaction settings this TikTok account permits",
        "tags": [
          "connections"
        ]
      }
    },
    "/connections/x": {
      "post": {
        "operationId": "ConnectionsController_connectX",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Connect X with an access token pair generated by hand, the redirect flow at x/authorize is the normal path",
        "tags": [
          "connections"
        ]
      }
    },
    "/connections/linkedin/authorize": {
      "get": {
        "operationId": "ConnectionsController_linkedinAuthorize",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the LinkedIn OAuth dialog URL",
        "tags": [
          "connections"
        ]
      }
    },
    "/connections/linkedin/callback": {
      "get": {
        "operationId": "ConnectionsController_linkedinCallback",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "OAuth redirect target for LinkedIn, do not call directly",
        "tags": [
          "connections"
        ]
      }
    },
    "/connections/tiktok/authorize": {
      "get": {
        "operationId": "ConnectionsController_tiktokAuthorize",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the TikTok OAuth dialog URL",
        "tags": [
          "connections"
        ]
      }
    },
    "/connections/tiktok/callback": {
      "get": {
        "operationId": "ConnectionsController_tiktokCallback",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "OAuth redirect target for TikTok, do not call directly",
        "tags": [
          "connections"
        ]
      }
    },
    "/connections/instagram/authorize": {
      "get": {
        "operationId": "ConnectionsController_instagramAuthorize",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the Instagram OAuth dialog URL (no Facebook Page required)",
        "tags": [
          "connections"
        ]
      }
    },
    "/connections/instagram/callback": {
      "get": {
        "operationId": "ConnectionsController_instagramCallback",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "OAuth redirect target for Instagram, do not call directly",
        "tags": [
          "connections"
        ]
      }
    },
    "/connections/meta/authorize": {
      "get": {
        "operationId": "ConnectionsController_metaAuthorize",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the Facebook OAuth dialog URL (Facebook Pages + Instagram)",
        "tags": [
          "connections"
        ]
      }
    },
    "/connections/meta/callback": {
      "get": {
        "operationId": "ConnectionsController_metaCallback",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "OAuth redirect target for Meta, do not call directly",
        "tags": [
          "connections"
        ]
      }
    },
...
  },
  "info": {
    "title": "GODEYE API",
    "description": "AI Marketing Operating System API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}